refactor: move builder screens into new subdirectory
