+++ apps/admin/frontend/src/screens/reporting/tally_report_builder.test.tsx
@@ -1,12 +1,12 @@
 import { electionTwoPartyPrimaryDefinition } from '@votingworks/fixtures';
 import userEvent from '@testing-library/user-event';
 import { expectPrint } from '@votingworks/test-utils';
-import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
-import { renderInAppContext } from '../../test/render_in_app_context';
+import { ApiMock, createApiMock } from '../../../test/helpers/mock_api_client';
+import { renderInAppContext } from '../../../test/render_in_app_context';
 import { TallyReportBuilder } from './tally_report_builder';
-import { screen, within } from '../../test/react_testing_library';
-import { getSimpleMockTallyResults } from '../../test/helpers/mock_results';
-import { canonicalizeFilter, canonicalizeGroupBy } from '../utils/reporting';
+import { screen, within } from '../../../test/react_testing_library';
+import { getSimpleMockTallyResults } from '../../../test/helpers/mock_results';
+import { canonicalizeFilter, canonicalizeGroupBy } from '../../utils/reporting';
 
 let apiMock: ApiMock;
